Introduce OapCacheLocations (#983)
This is a minor refactor, extract DAGScheduler.getOapCacheLocs method to a helper object named OapCacheLocations
