+++ src/main/spark2.3.2/scala/org/apache/spark/scheduler/DAGScheduler.scala
@@ -307,30 +307,6 @@ class DAGScheduler(
     cacheLocs(rdd.id)
   }
 
-  private [scheduler]
-  def getOapCacheLocs(rdd: RDD[_], partition: Int): Seq[TaskLocation] = {
-    val locations = rdd.preferredLocations(rdd.partitions(partition))
-
-    // here we check if the splits was cached, if the splits is cached, there will be hosts with
-    // format "OAP_HOST_host_OAP_EXECUTOR_exec", so we check the host list to filter out
-    // the cached hosts, so that the tasks' locality level will be PROCESS_LOCAL
-    if (locations.nonEmpty) {
-      // TODO use constant value for prefixes, these prefixes should be the same with that in
-      // [[org.apache.spark.sql.execution.datasources.oap.FiberSensor]]
-      val cacheLocs = locations.filter(_.startsWith("OAP_HOST_"))
-      val oapPrefs = cacheLocs.map { cacheLoc =>
-        val host = cacheLoc.split("_OAP_EXECUTOR_")(0).stripPrefix("OAP_HOST_")
-        val execId = cacheLoc.split("_OAP_EXECUTOR_")(1)
-        (host, execId)
-      }
-      if (oapPrefs.nonEmpty) {
-        logDebug(s"got oap prefer location value oapPrefs is ${oapPrefs}")
-        return oapPrefs.map(loc => TaskLocation(loc._1, loc._2))
-      }
-    }
-    Seq.empty
-  }
-
   private def clearCacheLocs(): Unit = cacheLocs.synchronized {
     cacheLocs.clear()
   }
@@ -1784,7 +1760,7 @@ class DAGScheduler(
     }
     // If the partition is cached, return the cache locations
     val cached = getCacheLocs(rdd)(partition)
-    val oapCached = getOapCacheLocs(rdd, partition)
+    val oapCached = OapCacheLocations.getOapCacheLocs(rdd, partition)
     if (cached.nonEmpty || oapCached.nonEmpty) {
       return cached ++ oapCached
     }
